:wrench: do not install if already there

+++ week1/setup.py
@@ -1,10 +1,12 @@
 import os
 
-# install virtualenv
-os.system('pip install virtualenv')
-# create virtualenv
-os.system('virtualenv venv')
-# activate the virtualenv
-os.system('source venv/bin/activate')
+if not os.path.exists('./venv'):
+    # install virtualenv
+    os.system('pip install virtualenv')
+    # create virtualenv
+    os.system('virtualenv venv')
+    # activate the virtualenv
+    os.system('source venv/bin/activate')
+
 # install python ext
 os.system('code --install-extension ms-python.python')
